ingest/requirements.txt
websocket-client==1.6.4
kafka-python==2.0.2
requests==2.31.0
prometheus-client==0.19.0
